+++ src/utils/constants.js
@@ -1,0 +1,3 @@
+module.exports.Constants = {
+  EMBED_COLOR: '#ff3860',
+}
